requirements.txt
python-docx
streamlit
pandas
scikit-learn
python-docx

